@@ -1,12 +1,12 @@
 class QiniuError extends Error {
-  final Error rawError;
+  final Error? rawError;
 
-  final String _message;
+  final String? _message;
 
   String get message => _message ?? rawError?.toString() ?? '';
 
   @override
-  StackTrace get stackTrace => rawError?.stackTrace ?? super.stackTrace;
+  StackTrace? get stackTrace => rawError?.stackTrace ?? super.stackTrace;
 
-  QiniuError({this.rawError, String message}) : _message = message;
+  QiniuError({this.rawError, String? message}) : _message = message;
 }
