@@ -22,19 +22,19 @@ class PutByPartTask extends RequestTask<PutResponse> {
   final int partSize;
   final int maxPartsRequestNumber;
 
-  final String key;
+  final String? key;
 
   /// 设置为 0，避免子任务重试失败后 [PutByPartTask] 继续重试
   @override
   int get retryLimit => 0;
 
   PutByPartTask({
-    @required this.file,
-    @required this.token,
-    @required this.partSize,
-    @required this.maxPartsRequestNumber,
+    required this.file,
+    required this.token,
+    required this.partSize,
+    required this.maxPartsRequestNumber,
     this.key,
-    PutController controller,
+    PutController? controller,
   })  : assert(file != null),
         assert(token != null),
         assert(partSize != null),
@@ -48,19 +48,16 @@ class PutByPartTask extends RequestTask<PutResponse> {
         }()),
         super(controller: controller);
 
-  RequestTaskController _currentWorkingTaskController;
+  RequestTaskController? _currentWorkingTaskController;
 
   @override
   void preStart() {
     super.preStart();
 
     // 处理相同任务
-    final sameTaskExsist = manager.getTasks().firstWhere(
-          (element) => element is PutByPartTask && isEquals(element),
-          orElse: () => null,
-        );
+    final sameTaskExsist = manager.getTasks().where((element) => element is PutByPartTask && isEquals(element)).toList();
 
-    if (sameTaskExsist != null) {
+    if (sameTaskExsist.isNotEmpty) {
       throw StorageError(
         type: StorageErrorType.IN_PROGRESS,
         message: '$file 已在上传队列中',
@@ -68,7 +65,7 @@ class PutByPartTask extends RequestTask<PutResponse> {
     }
 
     // controller 被取消后取消当前运行的子任务
-    controller?.cancelToken?.whenCancel?.then((_) {
+    controller?.cancelToken.whenCancel?.then((_) {
       _currentWorkingTaskController?.cancel();
     });
   }
