@@ -4,10 +4,10 @@ class PutBySingleOptions {
   /// 资源名
   ///
   /// 如果不传则后端自动生成
-  final String key;
+  final String? key;
 
   /// 控制器
-  final PutController controller;
+  final PutController? controller;
 
   const PutBySingleOptions({this.key, this.controller});
 }
